requirements.txt
pydub
numpy
openai-whisper